# Cache a few things to speed up canvas rendering.