Caching a few values and a scratch surface can give us some nice wins, particularly for rendering a zoomed in canvas.

## Original

|             Method |       Mean |     Error |    StdDev | Allocated |
|------------------- |-----------:|----------:|----------:|----------:|
|     RenderOneToOne |   2.159 ms | 0.0388 ms | 0.0363 ms |     242 B |
| RenderManyOneToOne |   9.597 ms | 0.1060 ms | 0.0940 ms |     824 B |
|       RenderZoomIn |  42.322 ms | 0.2513 ms | 0.2350 ms |  29,294 B |
|   RenderManyZoomIn | 404.774 ms | 3.2200 ms | 3.0120 ms |  29,600 B |
|      RenderZoomOut |   8.205 ms | 0.1055 ms | 0.0987 ms |     272 B |
|  RenderManyZoomOut |  76.499 ms | 0.7711 ms | 0.7213 ms |   1,574 B |

## PR

|             Method |       Mean |     Error |    StdDev | Allocated |
|------------------- |-----------:|----------:|----------:|----------:|
|     RenderOneToOne |   2.144 ms | 0.0155 ms | 0.0145 ms |     250 B |
| RenderManyOneToOne |   9.529 ms | 0.0866 ms | 0.0810 ms |     832 B |
|       RenderZoomIn |  22.584 ms | 0.1647 ms | 0.1460 ms |  28,358 B |
|   RenderManyZoomIn | 123.765 ms | 0.7484 ms | 0.7001 ms |  29,637 B |
|      RenderZoomOut |   8.179 ms | 0.0824 ms | 0.0771 ms |     280 B |
|  RenderManyZoomOut |  65.336 ms | 0.3871 ms | 0.3432 ms |     983 B |