@cameronwhite 2022-01-18
Thinking about this more, I wonder if the zoom in/out drawing could be done purely in Cairo by setting up a suitable transform for the source surface? The filtering behaviour would probably be different, but this could make it easy to draw straight onto the final surface with the layer's opacity / blend mode