@jpobst 2022-01-18
I was wondering about that, especially for 'Zoom Out'.  

I think 'Zoom In' is a bit trickier because most zoom functions apply a smoothing algorithm, and we want our 'Zoom In' to just show big, chunky pixels instead of being smoothed.