@cameronwhite 2022-01-18
Looking through the Cairo docs, it looks like that might be doable by setting up the "fast" or "nearest" filter for the source surface pattern